+++ src/redturtle/volto/upgrades.py
@@ -320,3 +320,46 @@ def to_2000(context):
             new_images.append(old_value)
 
     api.portal.set_registry_record("plone.allowed_sizes", new_images)
+
+
+def to_2100(context):  # noqa: C901
+    # reindex pages with table blocks
+
+    def has_table_block(blocks):
+        for block in blocks.values():
+            if block.get("@type", "") == "table":
+                return True
+        return False
+
+    pc = api.portal.get_tool(name="portal_catalog")
+    brains = pc()
+    tot = len(brains)
+    i = 0
+    for brain in brains:
+        i += 1
+        if i % 1000 == 0:
+            logger.info("Progress: {}/{}".format(i, tot))
+        item = aq_base(brain.getObject())
+        if getattr(item, "blocks", {}):
+            if has_table_block(item.blocks):
+                item.reindexObject(idxs=["SearchableText"])
+        for schema in iterSchemata(item):
+            # fix blocks in blocksfields
+            for name, field in getFields(schema).items():
+                if name == "blocks":
+                    blocks = getattr(item, "blocks", {})
+                    if has_table_block(blocks):
+                        item.reindexObject(idxs=["SearchableText"])
+                else:
+                    if not HAS_BLOCKSFIELD:
+                        # blocks are only in blocks field
+                        continue
+                    if isinstance(field, BlocksField):
+                        value = field.get(item)
+                        if not value:
+                            continue
+                        if isinstance(value, str):
+                            continue
+                        blocks = value.get("blocks", {})
+                        if has_table_block(blocks):
+                            item.reindexObject(idxs=["SearchableText"])